reducer overhaul: new style of reducers
Old reducers spliced a (map) body into a case-form to achieve
constant-time lookups & dispatching.
Unfortunately doing so required a macro. Because case itself is a
macro, it became impossible to splice-in the contents of a variable
- the reducer had to be supplied with the call to reducer-fn.

This leads to code duplication when reducers can't be composed
from smaller maps of action handlers (action maps henceforth).

This commit removes 'reducer-fn' in favour of 'reducer', a non-macro
which accepts one or more action maps, merges them together and
which will use lookups in the map as opposed to a case-form.
Otherwise, 'reducer' functions the same as before:
unhandled actions means the input model is returned & nil-models
are replaced by an initial value before the action is handled.